requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-dependency==0.5.1
pytest-env==0.8.1
python-dotenv==1.0.0


click~=8.1.3
Werkzeug~=2.2.3